+++ R/grid_search.R
@@ -22,8 +22,7 @@
 ##' @param sircovid_model An odin model generator and comparison function.
 ##'
 ##' @param model_params Model parameters, from a call to
-##'   \code{generate_parameters()}. If NULL, uses defaults as
-##'   in unit tests.
+##'   \code{generate_parameters()}
 ##'
 ##' @param pars_obs list of parameters to use for the comparison function.
 ##'
@@ -52,8 +51,8 @@ scan_beta_date <- function(
                            day_step,
                            data,
                            sircovid_model = basic_model(),
-                           model_params = NULL,
-                           pars_obs = NULL,
+                           model_params,
+                           pars_obs,
                            n_particles = 100,
                            scale_prior = NULL,
                            shape_prior = NULL,
@@ -76,34 +75,10 @@ scan_beta_date <- function(
   #
   # Set up calls to simulator runs
   #
-  if (is.null(model_params)) {
-    time_steps_per_day <- 4
-    model_params <- generate_parameters(
-      sircovid_model = sircovid_model,
-      transmission_model = "POLYMOD",
-      beta = 0.1,
-      beta_times = "2020-01-01",
-      trans_profile = 1,
-      trans_increase = 1,
-      dt = 1 / time_steps_per_day
-    )
-  } else {
     if (length(model_params$beta_y) > 1) {
       stop("Set beta variation through generate_beta_func in sircovid_model, not model_params")
     }
-  }
 
-  if (is.null(pars_obs)) {
-    pars_obs <- list(
-      phi_general = 0.95,
-      k_general = 2,
-      phi_ICU = 0.95,
-      k_ICU = 2,
-      phi_death = 926 / 1019,
-      k_death = 2,
-      exp_noise = 1e6
-    )
-  }
 
   #
   # Multi-core futures with furrr (parallel purrr)
@@ -187,56 +162,6 @@ plot.sircovid_scan <- function(x, ..., what = "likelihood", title = NULL) {
       x = x$x, y = x$y, z = x$renorm_mat_LL,
       xlab = "beta", ylab = "Start date", main = title
     )
-  }
-}
-
-##' @export
-plot.sample_grid_search <- function(x, ..., what = "ICU", title = NULL, col = 'grey80') {
-  idx <- odin_index(x$inputs$model$odin_model(
-    user = x$inputs$model_params,
-    unused_user_action = "message"
-  ))
-
-  # what are we plotting
-  if (what == "ICU") {
-    index <- c(idx$I_ICU_D, idx$I_ICU_R) - 1L
-    ylab <- "ICU"
-    particles <- vapply(seq_len(dim(x$trajectories)[3]), function(y) {
-      rowSums(x$trajectories[, index, y], na.rm = TRUE)
-    },
-    FUN.VALUE = numeric(dim(x$trajectories)[1])
-    )
-    plot_particles(particles, ylab = ylab, title = title, col = col)
-    points(as.Date(x$inputs$data$date), x$inputs$data$itu / x$inputs$pars_obs$phi_ICU, pch = 19)
-  } else if (what == "general") {
-    index <- c(idx$I_triage, idx$I_hosp_R, idx$I_hosp_D, idx$R_stepdown) - 1L
-    ylab <- "General beds"
-    particles <- vapply(seq_len(dim(x$trajectories)[3]), function(y) {
-      rowSums(x$trajectories[, index, y], na.rm = TRUE)
-    },
-    FUN.VALUE = numeric(dim(x$trajectories)[1])
-    )
-    plot_particles(particles, ylab = ylab, title = title, col = col)
-    points(as.Date(x$inputs$data$date), x$inputs$data$general / x$inputs$pars_obs$phi_general, pch = 19)
-  }
-
-  else if (what == "deaths") {
-    index <- c(idx$D) - 1L
-    ylab <- "Deaths"
-    particles <- vapply(seq_len(dim(x$trajectories)[3]), function(y) {
-      out <- c(0, diff(rowSums(x$trajectories[, index, y], na.rm = TRUE)))
-      names(out)[1] <- rownames(x$trajectories)[1]
-      out
-    },
-    FUN.VALUE = numeric(dim(x$trajectories)[1])
-    )
-    plot_particles(particles, ylab = ylab, title = title, col = col)
-    points(as.Date(x$inputs$data$date),
-      x$inputs$data$deaths / x$inputs$pars_obs$phi_death,
-      pch = 19
-    )
-  } else {
-    stop("Requested what must be one of 'ICU', 'deaths' or 'general'")
   }
 }